+++ src/index.c
@@ -4,10 +4,10 @@
 int init_index(const struct bgen_file *bgen, struct bgen_vi *vi)
 {
     vi->filepath = strdup(bgen->filepath);
-    vi->compression = bgen->compression;
-    vi->layout = bgen->layout;
-    vi->nsamples = bgen->nsamples;
-    vi->nvariants = bgen->nvariants;
+    vi->compression = (uint32_t)bgen->compression;
+    vi->layout = (uint32_t)bgen->layout;
+    vi->nsamples = (uint32_t)bgen->nsamples;
+    vi->nvariants = (uint32_t)bgen->nvariants;
     vi->max_nalleles = 0;
     /* `start` is not used anymore. */
     vi->start = NULL;
